# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for maintenance of SWIFTcode API.",
        "title": "SWIFTcode API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/swift-codes": {
            "get": {
                "description": "Retrieves all headquarters (without their subsequent branches) and all branches from the SWIFT database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "swift-codes headquarter branch"
                ],
                "summary": "Get all SWIFT entities (headquarters and branches)",
                "responses": {
                    "200": {
                        "description": "List of SWIFT entries",
                        "schema": {
                            "$ref": "#/definitions/REST.AllResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/REST.MessageResponse"
                        }
                    },
                    "404": {
                        "description": "No entries found",
                        "schema": {
                            "$ref": "#/definitions/REST.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/REST.MessageResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Add new headquarter or branch to the SWIFT database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "headquarter branch"
                ],
                "summary": "Add a new entry",
                "parameters": [
                    {
                        "description": "Entry data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/REST.BranchResponse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful addition",
                        "schema": {
                            "$ref": "#/definitions/REST.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/REST.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/REST.MessageResponse"
                        }
                    }
                }
            }
        },
        "/swift-codes/country/{ISO2}": {
            "get": {
                "description": "Retrieves a country's name, ISO2 and its subsequent headquarters (without the subsequent branches) and branches from the SWIFT database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "country"
                ],
                "summary": "Get data about the country and its subsequent entries",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"PL\"",
                        "description": "ISO2",
                        "name": "ISO2",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Country exists",
                        "schema": {
                            "$ref": "#/definitions/REST.CountryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/REST.MessageResponse"
                        }
                    },
                    "404": {
                        "description": "No country found",
                        "schema": {
                            "$ref": "#/definitions/REST.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/REST.MessageResponse"
                        }
                    }
                }
            }
        },
        "/swift-codes/{swift-code}": {
            "get": {
                "description": "Returns the information about the headquarter (with subsequent branches) or the branch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "swift-codes"
                ],
                "summary": "Get SWIFT entity",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"MRWORLDWXXX\"",
                        "description": "SWIFT code",
                        "name": "swift-code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SWIFT entity",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a headquarter (without the subsequent branches) or branch specified by the SWIFT code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "headquarter branch"
                ],
                "summary": "Delete a SWIFT determined entry",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"MRWORLDWXXX\"",
                        "description": "SWIFT code",
                        "name": "swift-code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entry successfully deleted",
                        "schema": {
                            "$ref": "#/definitions/REST.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/REST.MessageResponse"
                        }
                    },
                    "404": {
                        "description": "No entry found",
                        "schema": {
                            "$ref": "#/definitions/REST.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/REST.MessageResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "REST.AllResponse": {
            "type": "object",
            "properties": {
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/structs.ReqHeadBranInCountry"
                    }
                }
            }
        },
        "REST.BranchResponse": {
            "type": "object",
            "required": [
                "bankName",
                "countryISO2",
                "countryName",
                "isHeadquarter",
                "swiftCode"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "maxLength": 255
                },
                "bankName": {
                    "type": "string",
                    "maxLength": 255
                },
                "countryISO2": {
                    "type": "string",
                    "maxLength": 2,
                    "minLength": 2
                },
                "countryName": {
                    "type": "string",
                    "maxLength": 20
                },
                "isHeadquarter": {
                    "type": "boolean"
                },
                "swiftCode": {
                    "type": "string",
                    "maxLength": 11,
                    "minLength": 11
                }
            }
        },
        "REST.CountryResponse": {
            "type": "object",
            "properties": {
                "countryISO2": {
                    "type": "string"
                },
                "countryName": {
                    "type": "string"
                },
                "swiftCodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/structs.ReqHeadBranInCountry"
                    }
                }
            }
        },
        "REST.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "structs.ReqHeadBranInCountry": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "bankName": {
                    "type": "string"
                },
                "countryISO2": {
                    "type": "string"
                },
                "isHeadquarter": {
                    "type": "boolean"
                },
                "swiftCode": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}